requirements.txt
torch==2.0.1
numpy>=1.21.0
Pillow>=9.0.0
opencv-python
requests